Reimplemented Background and CorrectedImage calculations
Background
==========
sklearn.svm.SVR.predict took a long time to predict the full background.
Instead, I'm now predicting a subset of the background on a regular grid,
and then doing an spline interpolation.
It decreased runtime by a factor x50, from ~1 min to ~1 sec.
Also, it decreased file output size by x1000.

CorrectedImage
==============
I'm not sure the correct way to substract the background and
apply the normalization. For now, I'm ignoring that correction.

Signed-off-by: Mauro Silberberg <maurosilber@gmail.com>