Updated Intensities to include more cells
Previously, saturated cells were filled with NaN,
which propagated and excluded that data point from further analysis.
Now, saturated areas are excluded from analysis,
and the intensity is integrated in non-saturated areas.

Additionally, measures of saturated area are also extracted.

Also, cells which touched the border of the image were excluded from further analysis.
Now, there is a boolean array which signals if a cell is touching the border.

Signed-off-by: Mauro Silberberg <maurosilber@gmail.com>